Add hidden service sorter and central config file
The hidden service sorter scrapes the SecureDrop directory as well as ahmia.fi
to look for hidden service URLs. (Alternate directories to scrape may be
configured in the config.ini file.) It then scrapes these hidden services in
order to sort them into four categories: into up-to-date & deprecated SDs, sites
that mention SD, and sites that are neither SD or mention SD. Separating SDs
from non-SDs in the HS space is necessary because there may be SDs that we do
not know about or otherwise do not have listed in the securedrop.org directory.