Convert url sets to list before pickling
With the new crawler design intended to be pushed by tomorrow, raw tor logging
data will be immediately processed as follows:

Each site will be identified with it's index (such that without access to the
pickled data, the mapping may between the index and site may remain unknown).
Each trace will be saved to a file named as follows: <index>-<iteration>, where
<iteration> refers to the ith iterative attempt to crawl this particular site in
this particular run of the crawler.  Each trace will be a recorded as a sequence
of 2-tuples of the form '<timestamp> <direction>', each 2-tuple being
newline-delimited from the previous.

The reason for the change in format is three-fold: (i) it allows us to anonymize
our dataset, (ii) it gives a static identifier to each site such that data from
different crawls can be combined, (iii) it allows us to directly analyze the
output of the crawler with go-knn.