Main loop reorg: close circs in `finally:`, `continue` after weird excs
This commit is a mild reorganization of the main loop that collects traces from
all the URLs in a class. It adds a missing `continue` statement to the block
that is a catch-all for the more esoteric exceptions encountered. Also, it moves
the operation to close all open circuits to the `finally` block in this loop, so
that this cleanup always happens.

Assume the last onion service partially loaded, but timed out (just an
example--this could apply to any exceptions in this loop before this change).
While Tor Browser should close the circuit to the last onion service it tried to
reach as soon as we try to load the new one (since the new site is loaded in the
same tab), closing all circuits in finally eliminates the possibility that a
couple cells from the old circuit make it into the trace for the new site. They
could sneak in between when we seek to the end of the Tor cell log and when Tor
actually kills the old circuit.