Supports --tags and --skip-tags via ANSIBLE_ARGS env var
Since we're using core Vagrant for managing VMs, and Ansible for
provisioning, we'll need to hook Ansible into Vagrant to support ad-hoc
changes for individual playbooks runs, e.g. `--tags` and `--skip-tags`.
It's a bit of a hack to do this via an environment variable, but
switching the VM/test suite to a more modern Ansible-friendly framework
such as Molecule seems heavy-handed. This works well enough.

Besides, in prod there's a whole other playbook in the mix anyway, and
prod deployments are done (at least in the FPF environment) via
`ansible-playbook` anyway, so raw CLI options are already available
there.