Add variable for pause between circuit closing and next trace
We might be closing certain circuits which will just be rebuilt pretty
immediately regardless of the fact we're immediately thereafter loading an onion
service. That could cloud how the beginning of a trace on a regular system might
look.

In Wang's work he pauses 5s after closing circuits before crawling a new one.
We're going to experiment with different values.