Makes tor source and tbb downloads idempotent
Reenables certificate validation, which is a Good Idea. The trick to
achieving idempotence with the Ansible `get_url` module is that you must
specify a filename for the file, otherwise Ansible can't be sure the
file has already been downloaded. We can create a reasonable filename by
using the 'basename' filter to shave off the leading part of the URL.

Ansible won't redownload the file if the filename exists. It also won't
be able to detect incomplete or interrupted downloads, but we're
verifying integrity via GPG signatures in subsequent tasks.