Ensure symlinks are removed even when broken
"It's easier to ask forgiveness than it is to get permission."

os.stat() tries to follow broken symlinks and then reports the file they point
to as missing, rather than letting us know the symlink itself exists (even if
it is broken). This would create problems periodically if log files were
deleted or moved, so now we just try to remove the symlink first, and then
catch the exception if it doesn't exist.