Ensure pip updates Python packages when versions change
If a package is already installed to the target environment, pip will not
upgrade/ downgrade that package (even if a different version is requested than
is already installed) unless the -U/--upgrade option is passed. I confirmed in
the Ansible pip module source code that specifying the state=latest parameter
does pass the -U option when installing packages.