Use /etc/bash.bashrc to set env vars instead of /etc/profile.d/...
So /etc/profile and /etc/profile.d/* are only sourced by login shells. So you
need to do sudo -Es && su postgres to keep the env vars if you need to operate
as the postgres user with the way things are set up currently. Since we
shouldn't need to support other shells, it's probably better that these are set
in /etc/bash.bashrc so that they're always accessible to all bash users.