Create test database for use w/ test fixtures
Create a test database with the same schemas and tables used in the real one.
This is accomplished by simply creating the test database at the end of
provisioning, and choosing the now-fully-initialized "fpsd" database as a
template.

The test database should be wiped after each use by the test fixtures, in the
tearDown() method. Hence I have introduced the wipe_database method, which, for
safety reasons, refuses to operate on database if its name starts with "test."
This method also gets tested in the database methods test during the tearDown().