Adds support for `--check` mode to database role
Ansible offers the `--check` flag for performing dry-runs, so admins can
get a sense of what specifically would change on a target host prior to
applying the role for real. By setting always_run=true on tasks that
register vars, we can ensure that the registered vars are at least
defined, which allows subsequent tasks referencing the registered vars
to run successfully (or skip, as if often the case).

The --check mode support won't work during the first-run of the role,
but will work for every subsequent change after that, which is good
enough.