Ensure Sorter class_tests order is respected
I mistakenly believe `collections.OrderedDict`, when initialized with multiple
entries, would respect the written order of those entries. Contrarily, it does
not, and in order to preserve that order we need to add on entry at a time.

I chose to use %% as a delimiter simply because I figured it's unlikely to come
up in any boolean expression a user might want to use.

Fixes #33.