Updates the logic to infer a username for configuring dotfiles
We can't assume that `ansible_ssh_user` will be defined, since it's not
populated by default in Ansible. When provisioning Vagrant VMs, the
`ansible_ssh_user` var is defined because it's hardcoded in the static
inventory file that Vagrant generates.

Let's prefer `ansible_user`, which supersedes `ansible_ssh_user` as of v2,
but still honor `ansible_ssh_user` to support Vagrant. If neither can be
found, we'll fall back to using the username of the account running
ansible on the Ansible controller.