Folds common and sorter roles into crawler role
The sorter role had a single task in it, installing Python dependencies
via pip from a requirements.txt file. The location of the
requirements.txt was determined by a previous role, which is dirty, so
let's merge the task lists so we can reuse vars.

The apt packages, pip dependencies, and git cloning tasks were
previously housed in the `common` role, and are now in the crawler role
where they belong.