More durable changed status on tor-src dependencies
The logic for the deb-src parameter in the apt module inappropriately
reports changed=yes always, so let's inspect the output and report
accordingly. Previously the inspection mandated that all packages were
fully up to date, but that's not an accurate representation of changed
status for this task: only if packages were installed or upgraded should
the task report changed.