Makes crawler service configuration optional
By default the role will still apply the crawler code tasks, but it's
now possible to skip them via a var in order to run just the database or
credential config tasks. This allows the same role and vars setup to be
distributing across multiple hosts when running the crawler in
production.