Fix EntryNodes line in torrc
When you launch tor with stem, you may specify torrc options via a dict. If you
pass an iterable as a value for any key in that dict, it will create a separate
line that begins with the key str and ends with a str in that iterable, for each
iterable (the Cartesian product of the key and each val in the iterable).

I've made a tiny change to config.ini + sorter.ini to standardize delimiters in
the file a bit accordingly.

CookieAuth has been deprecated in favor of CookieAuthentication.