Fetch db config from config.ini
I thought it was best the code stopped reading from environment variables since
it was causing problems with Travis. To do this, I implemented a base class,
Database, and a mixin class TestDatabase. Using cooperative multiple
inheritance, one can easily create a test class for any ancestor of Database,
which is exactly what is now occuring in the tests.