Simplifies group membership update task
The original intent here is to add all authorized users to the `fpsd`
group so that developers can run commands as the `fpsd` service account
without directly su'ing as the service account user. Makes sense.

The previous implementation was using a two-pass approach to accommodate
for improper idempotence handling in Ansible's `user` module. Was able
to reproduce the issue reliably on Ansible v2.1.1, but Ansible v2.2.0
has resolved the issue, and the role applies cleanly on multiple runs
with the new one-task approach.